v2.5 overhaul, broke up the ck script
