+++ scripts/request_wrapper.py
@@ -9,6 +9,7 @@ def _send_request(method:str, url:str, **kwargs) -> dict:
     """
     Catch all wrapper for sending web request and interpreting as json
     """
+
     log.info(f"Sending request to {url} with args {kwargs}")
     r = requests.request(method, url, **kwargs)
 
@@ -21,18 +22,5 @@ def _send_request(method:str, url:str, **kwargs) -> dict:
     try:
         data = json.loads(r.text)
     except json.JSONDecodeError as e:
-        log.exception("Failed to decode response as json, was there an error in request?: {e}")
-
+        log.exception(f"Failed to decode response as json, was there an error in request?: {e}")
     return data
-
-
-headers = {
-    'user-agent': 'Mozilla 5.0/Chrome'  # Simple anti bot bypass
-}
-# cookies = CookeiJar(...)
-# r = _send_request('GET', 'https://httpbin.org/get', headers=headers, cookies=cookies)
-# print(r)
-
-requests.se
-
-
